# Conflicts:
#	src/main/java/my_project/control/CookingController.java

+++ src/main/java/my_project/model/GUI/HPBar.java
@@ -2,11 +2,13 @@ package my_project.model.GUI;
 
 import KAGO_framework.view.DrawTool;
 import my_project.model.Environment.CollidableEnvironment;
-import my_project.model.UI;
 
+/**
+ * Displays the HP of an attached CollidableEnvironment object
+ */
 public class HPBar extends UI {
 
-    private double health;
+    private int health;
     private CollidableEnvironment environment;
 
     /**
